pre-setup: simplify uname's and extend installer
The checker for the OS is simpler, the usage of `uname -o | grep -q -i
linux` catches a lot more possibilities than plain Linux. For example,
in voidlinux with glibc it is "GNU/Linux" which wouldn't be matched with
the former `[ "$(uname -o)" = "Linux" ]`.

The installer of `rcm` is extended and allows the possibility of
installing on linux with apk and apt if any of those commands exists. It
won't default to xbps if the command does not exist. If it doesn't knows
what to do then it'll fail asking to please install rcm pointing to its
GitHub repository.